+++ NEWS.md
@@ -1,5 +1,16 @@
 ### HLAtools
 
+### version 1.5.0
+- April 20, 2025
+- Refactored the buildIMGTHLAGeneTypes() function to accommodate the transition of the "molecular characteristics data for HLA region genes" web-page from 'hla.alleles.org/genes/index.html' to 'hla.alleles.org/pages/genes/genes_list'.
+- Updated updateAll() to account for the transition from 'hla.alleles.org/genes/index.html' to 'hla.alleles.org/pages/genes/genes_list' for updating IMGTHLAGeneTypes.
+- Added the validatePositions() function to determine if specified positions exist in a specified type of alignment. 
+- Applied validatePosition() for customAlign(), and updated the customAlign() documentation.
+- Updated buildGazeteer() to include the DRB8 gene in HLAgazeteer$align.
+- Updated IMGTHLAGeneTypes, HLAgazeteer and HLAtlas to version 3.60.0.
+- Added the rvest and stats packages to Imports.
+- The alleleListHistory dataset for IPD-IMGT/HLA Database release versions 3.60.0 and later is too large to be included in the package. The updateAll() function can be applied to load the alleleListHistory object for releases after 3.59.0.
+
 ### Version 1.4.0
 
 - March 1, 2025